requirements.txt
asgiref==3.5.0
asttokens==2.0.5
autopep8==1.6.0
backcall==0.2.0
backports.zoneinfo==0.2.1
debugpy==1.5.1
decorator==5.1.1
Django==4.0.3
django-environ==0.8.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
entrypoints==0.4
executing==0.8.3
ipykernel==6.9.1
ipython==8.1.1
jedi==0.18.1
jupyter-client==7.1.2
jupyter-core==4.9.2
matplotlib-inline==0.1.3
nest-asyncio==1.5.4
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.28
ptyprocess==0.7.0
pure-eval==0.2.2
pycodestyle==2.8.0
Pygments==2.11.2
PyJWT==2.3.0
python-dateutil==2.8.2
pytz==2021.3
pyzmq==22.3.0
six==1.16.0
sqlparse==0.4.2
stack-data==0.2.0
toml==0.10.2
tornado==6.1
traitlets==5.1.1
wcwidth==0.2.5
